lightbulb_bot/requirements.txt
discord.py
discord.py[voice]
yt-dlp
PyNaCl
asyncio
easy-pil
torch
diffusers
transformers
accelerate